Add files via upload
Day 24 - Started to work with Files, Directories and Paths.  In this main.py program, I had an input folder which includes names of letter recipients and a draft letter, as well as an output folder to store the ready to send letters.  I created an automation that created individual letters for each person on the list of recipients, reducing redundancy in letter writing and saving time.