Add files via upload
Day 25 - I started to work with CSV Data and the Pandas Library.  I created programs to practice sifting through data.  Weather_data.py was my intro to Pandas.  In Squirrel_main.py, I sifted through the 2018 Central Park Squirrel count to get an estimate of how many squirrels of each color there were in the given year.  In stategame.py, I created a state guessing game learning tool that took a user input and compared to the list of states.  Once guesses are made correctly, the name of the state pops up on the image.  If player chooses to exit game prematurely, a CSV file of the missing states is created for study purposes.