Add files via upload
Day 26 - Learned about list and dictionary comprehensions, practiced a variety of comprehensions.  In this upload, I created a dictionary from the attached csv using Pandas.  From there, I created a list comprehension that references the new dictionary.  So, a user enters any word and the computer responds with a phonetic list.