Add files via upload
Day 28 - Created a GUI app that serves as a time management/productivity tool.  The GUI runs as a timer for 25 minutes before giving a five-minute break.  On the fourth go, give a 20-minute break.  Also keeps track of how many work sessions have been completed since pressing start and shows in the form of check marks.